relocation services,interfaces and components
