+++ src/app/app-routing.module.ts
@@ -1,13 +1,10 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { PageNotFoundComponent } from './components/page-not-found/page-not-found.component';
+import { PageNotFoundComponent } from './shared/page-not-found/page-not-found.component';
 
 const routes: Routes = [
   { path: '', loadChildren: () => import('./home/home.module').then ( m => m.HomeModule) },
   { path: 'auth', loadChildren: () => import('./auth/auth.module').then( m => m.AuthModule) },
-  // {path: 'login', component: LoginComponent},
-  // {path: 'registrarse', component: RegisterComponent},
-  // {path: 'configuracion', component: ConfigurationComponent},
   {path: '**', component: PageNotFoundComponent}
 ];
