+++ src/app/app.module.ts
@@ -8,78 +8,30 @@ import {AngularFirestoreModule, SETTINGS as FIRESTORE_SETTINGS} from '@angular/f
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { HeaderComponent } from './components/header/header.component';
-import { ConfigurationComponent } from './components/configuration/configuration.component';
-import { PageNotFoundComponent } from './components/page-not-found/page-not-found.component';
-import { FooterComponent } from './components/footer/footer.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 
-import {MatToolbarModule} from '@angular/material/toolbar';
-import {MatIconModule} from '@angular/material/icon';
-import {MatButtonModule} from '@angular/material/button';
-import {MatSidenavModule} from '@angular/material/sidenav';
-import {MatListModule} from '@angular/material/list';
-import {MatExpansionModule} from '@angular/material/expansion';
-import {MatDividerModule} from '@angular/material/divider';
-import {MatMenuModule} from '@angular/material/menu';
-import {MatTableModule} from '@angular/material/table';
-import {MatCardModule} from '@angular/material/card';
-import { ClientsServices } from './services/clients.service';
-import { MatFormFieldModule } from '@angular/material/form-field';
-import {MatInputModule} from '@angular/material/input';
 import {MatProgressSpinnerModule} from '@angular/material/progress-spinner';
-import { LoginService } from './services/login.service';
-import { DialogAddClientComponent } from './components/dialog-add-client/dialog-add-client.component';
 import { MatDialogModule } from '@angular/material/dialog';
-import { MatPaginatorModule } from '@angular/material/paginator';
-import {MatCheckboxModule} from '@angular/material/checkbox';
-import { ConfigurationService } from './services/configuration.service';
 import { MatSnackBarModule } from '@angular/material/snack-bar';
-import { ConfirmationDialogComponent } from './components/confirmation-dialog/confirmation-dialog.component';
-import { SnackBarService } from './services/snackBar.service';
-import { DialogSendEmailComponent } from './components/dialog-send-email/dialog-send-email.component';
+
 @NgModule({
   declarations: [
     AppComponent,
-    HeaderComponent,
-    ConfigurationComponent,
-    PageNotFoundComponent,
-    FooterComponent,
-    DialogAddClientComponent,
-    ConfirmationDialogComponent,
-    DialogSendEmailComponent,
   ],
   imports: [
     BrowserModule,
     AppRoutingModule,
-    MatToolbarModule,
     BrowserAnimationsModule,
+    FormsModule,
     AngularFireModule.initializeApp(environment.firebaseConfig, 'control-clientes'),
     AngularFirestoreModule,
-    MatIconModule,
-    MatPaginatorModule,
-    MatButtonModule,
-    MatSidenavModule,
-    MatListModule,
-    MatExpansionModule,
-    MatDividerModule,
-    MatMenuModule,
-    MatTableModule,
-    MatCardModule,
-    MatFormFieldModule,
-    MatInputModule,
-    FormsModule,
     ReactiveFormsModule,
     MatProgressSpinnerModule,
     MatDialogModule,
-    MatCheckboxModule,
     MatSnackBarModule
   ],
   providers: [
-    ClientsServices,
-    LoginService,
-    ConfigurationService,
-    SnackBarService,
+
     {provide: FIRESTORE_SETTINGS, useValue:{}}
   ],
   bootstrap: [AppComponent],
