+++ src/app/auth/components/dialog-send-email/dialog-send-email.component.ts
