+++ src/app/core/interfaces/configuration.interface.ts
