+++ src/app/core/interfaces/error-type.interface.ts
