+++ src/app/home/home.module.ts
@@ -13,6 +13,11 @@ import { MatTableModule } from '@angular/material/table';
 import { MatPaginatorModule } from '@angular/material/paginator';
 import { MatCardModule } from '@angular/material/card';
 import { MatButtonModule } from '@angular/material/button';
+import { ConfigurationComponent } from './pages/configuration/configuration.component';
+import { MatCheckboxModule } from '@angular/material/checkbox';
+import { SharedModule } from '../shared/shared.module';
+import { DialogAddClientComponent } from './components/dialog-add-client/dialog-add-client.component';
+import { ConfirmationDialogComponent } from './components/confirmation-dialog/confirmation-dialog.component';
 
 
 @NgModule({
@@ -20,12 +25,16 @@ import { MatButtonModule } from '@angular/material/button';
     ClientsComponent,
     DashboardComponent,
     HomeLayoutComponent,
+    ConfigurationComponent,
+    DialogAddClientComponent,
+    ConfirmationDialogComponent,
   ],
   imports: [
     CommonModule,
     FormsModule,
     HomeRoutingModule,
     ReactiveFormsModule,
+    SharedModule,
 
     MatButtonModule,
     MatCardModule,
@@ -34,6 +43,7 @@ import { MatButtonModule } from '@angular/material/button';
     MatInputModule,
     MatPaginatorModule,
     MatTableModule,
+    MatCheckboxModule,
   ]
 })
 export class HomeModule { }
