+++ src/app/home/layouts/home-layout/home-layout.component.html
@@ -1,1 +1,6 @@
+<app-header></app-header>
+
 <router-outlet></router-outlet>
+
+<!--TODO: aplicar estilos al footer-->
+<!-- <app-footer></app-footer> -->
