+++ src/app/home/pages/configuration/configuration.component.css
