+++ src/app/home/pages/configuration/configuration.component.html
@@ -8,7 +8,7 @@
 </header>
 
 <section class="buttons">
-    <a mat-button routerLink="/login">
+    <a mat-button routerLink="/home">
         <mat-icon>arrow_back</mat-icon>
         Regresar Al Inicio
     </a>
