+++ src/app/shared/shared.module.ts
@@ -1,0 +1,31 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { HeaderComponent } from './header/header.component';
+import { MatIconModule } from '@angular/material/icon';
+import { MatButtonModule } from '@angular/material/button';
+import { MatMenuModule } from '@angular/material/menu';
+import { MatToolbarModule } from '@angular/material/toolbar';
+import { PageNotFoundComponent } from './page-not-found/page-not-found.component';
+import { FooterComponent } from './footer/footer.component';
+
+
+
+@NgModule({
+  declarations: [
+    HeaderComponent,
+    PageNotFoundComponent,
+    FooterComponent,
+  ],
+  imports: [
+    CommonModule,
+    MatIconModule,
+    MatButtonModule,
+    MatMenuModule,
+    MatToolbarModule,
+  ],
+  exports: [
+    HeaderComponent,
+    FooterComponent,
+  ]
+})
+export class SharedModule { }
